+++ middlewares/validation.js
@@ -1,0 +1,21 @@
+const Joi = require("joi");
+
+module.exports = {
+  schemaValidation: (req, res, next) => {
+    const schema = Joi.object({
+      name: Joi.string().max(30).alphanum().required(),
+
+      phone: Joi.string().max(12).required(),
+
+      email: Joi.string().email({
+        minDomainSegments: 2,
+        tlds: { allow: ["com", "net"] },
+      }),
+    });
+    const validation = schema.validate(req.body);
+    if (validation.error) {
+      return res.status(400).json({ message: "missing required name field" });
+    }
+    next();
+  },
+};
